// web/src/api/models/ConfigLLDPConfig.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ConfigLLDPConfig
 */
export interface ConfigLLDPConfig {
    /**
     * 
     * @type {boolean}
     * @memberof ConfigLLDPConfig
     */
    cdp?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ConfigLLDPConfig
     */
    enabled?: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof ConfigLLDPConfig
     */
    interfaces?: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof ConfigLLDPConfig
     */
    transmit?: boolean;
}

/**
 * Check if a given object implements the ConfigLLDPConfig interface.
 */
export function instanceOfConfigLLDPConfig(value: object): value is ConfigLLDPConfig {
    return true;
}

export function ConfigLLDPConfigFromJSON(json: any): ConfigLLDPConfig {
    return ConfigLLDPConfigFromJSONTyped(json, false);
}

export function ConfigLLDPConfigFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConfigLLDPConfig {
    if (json == null) {
        return json;
    }
    return {
        
        'cdp': json['cdp'] == null ? undefined : json['cdp'],
        'enabled': json['enabled'] == null ? undefined : json['enabled'],
        'interfaces': json['interfaces'] == null ? undefined : json['interfaces'],
        'transmit': json['transmit'] == null ? undefined : json['transmit'],
    };
}

export function ConfigLLDPConfigToJSON(json: any): ConfigLLDPConfig {
    return ConfigLLDPConfigToJSONTyped(json, false);
}

export function ConfigLLDPConfigToJSONTyped(value?: ConfigLLDPConfig | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'cdp': value['cdp'],
        'enabled': value['enabled'],
        'interfaces': value['interfaces'],
        'transmit': value['transmit'],
    };
}